Change hashing to prevent collisions.
Previously, having albums or photos with the same names could result
in serving incorrect files. This was a security risk for people who
use the same naming scheme in every album.

Also use a MD5 instead of a SHA1 since it's shorter while still
sufficient to prevent collisions and enumeration attacks.